Support different versions of schema tables NODEJS-163
