+++ lib/control-connection.js
@@ -5,18 +5,16 @@ var util = require('util');
 var dns = require('dns');
 var net = require('net');
 
-var errors = require('./errors.js');
-var RequestHandler = require('./request-handler.js');
-var Host = require('./host.js').Host;
-var HostMap = require('./host.js').HostMap;
-var Metadata = require('./metadata.js');
-var requests = require('./requests.js');
-var utils = require('./utils.js');
+var errors = require('./errors');
+var RequestHandler = require('./request-handler');
+var Host = require('./host').Host;
+var HostMap = require('./host').HostMap;
+var Metadata = require('./metadata');
+var requests = require('./requests');
+var utils = require('./utils');
 
 var selectPeers = "SELECT peer,data_center,rack,tokens,rpc_address,release_version FROM system.peers";
 var selectLocal = "SELECT * FROM system.local WHERE key='local'";
-var selectAllKeyspaces = "SELECT * FROM system.schema_keyspaces";
-var selectSingleKeyspace = "SELECT * FROM system.schema_keyspaces where keyspace_name = '%s'";
 var newNodeDelay = 1000;
 var retryNewConnectionDelay = 5000;
 var schemaChangeTypes = {
@@ -132,6 +130,7 @@ ControlConnection.prototype.getConnection = function (callback) {
     if (!err) {
       self.connection = c;
       self.host = host;
+      self.metadata.setCassandraVersion(host.getCassandraVersion());
     }
     callback(err);
   }
@@ -156,7 +155,7 @@ ControlConnection.prototype.initOnConnection = function (firstTime, callback) {
       self.refreshHosts(firstTime, next);
     },
     function getKeyspaces(next) {
-      self.getKeyspaces(next);
+      self.metadata.refreshKeyspaces(next);
     },
     function buildTokenPerReplica(next) {
       if (self.metadata.tokenizer) {
@@ -233,16 +232,6 @@ ControlConnection.prototype.hostDownHandler = function () {
       }
     }
   );
-};
-
-ControlConnection.prototype.getKeyspaces = function (callback) {
-  this.log('info', 'Retrieving keyspaces metadata');
-  var self = this;
-  this.query(selectAllKeyspaces, function (err, result) {
-    if (err) return callback(err);
-    self.metadata.setKeyspaces(result);
-    callback();
-  });
 };
 
 /**
@@ -324,19 +313,7 @@ ControlConnection.prototype.nodeSchemaChangeHandler = function (event) {
     }
   }
   //The keyspace was either created or altered
-  var query = util.format(selectSingleKeyspace, event.keyspace);
-  var self = this;
-  this.query(query, function (err, result) {
-    if (err) {
-      self.log('error', 'There was an error while trying to retrieve keyspace information', err);
-      return;
-    }
-    if (result.rows.length === 0) {
-      self.log('warning', 'There was not possible to retrieve keyspace info', event.keyspace);
-      return;
-    }
-    self.metadata.setKeyspaceInfo(result.rows[0]);
-  });
+  this.metadata.refreshKeyspace(event.keyspace);
 };
 
 ControlConnection.prototype.setLocalInfo = function (endPoint, result) {
