+++ lib/host.js
@@ -23,7 +23,7 @@ function Host(address, protocolVersion, options) {
   Object.defineProperty(this, 'pool', { value: new HostConnectionPool(address, protocolVersion, options), enumerable: false});
   this.pool.on('idleRequestError', this.setDown.bind(this));
   /**
-   * Gets Cassandra version string.
+   * Gets string containing the Cassandra version.
    * @type {String}
    */
   this.cassandraVersion = null;
@@ -171,6 +171,18 @@ Host.prototype.getActiveConnection = function () {
  */
 Host.prototype.checkHealth = function (connection) {
   this.pool.checkHealth(connection);
+};
+
+/**
+ * Returns an array containing the Cassandra Version as an Array of Numbers having the major version in the first
+ * position.
+ * @returns {Array.<Number>}
+ */
+Host.prototype.getCassandraVersion = function () {
+  if (!this.cassandraVersion) return utils.emptyArray;
+  return this.cassandraVersion.split('-')[0].split('.').map(function (x) {
+    return parseInt(x, 10);
+  });
 };
 
 Host.prototype.log = utils.log;
