+++ lib/metadata/table-metadata.js
@@ -1,0 +1,109 @@
+"use strict";
+var util = require('util');
+var events = require('events');
+/** @module metadata */
+/**
+ * Creates a new instance of TableMetadata
+ * @param {String} name
+ * @class
+ * @classdesc Describes a table
+ * @constructor
+ */
+function TableMetadata(name) {
+  events.EventEmitter.call(this);
+  this.setMaxListeners(0);
+  //private
+  Object.defineProperty(this, 'loading', { value: false, enumerable: false, writable: true });
+  Object.defineProperty(this, 'loaded', { value: false, enumerable: false, writable: true });
+  /**
+   * Name of the table
+   * @type {String}
+   */
+  this.name = name;
+  /**
+   * False-positive probability for SSTable Bloom filters.
+   * @type {number}
+   */
+  this.bloomFilterFalsePositiveChance = 0;
+  /**
+   * Level of caching: all, keys_only, rows_only, none
+   * @type {String}
+   */
+  this.caching = null;
+  /**
+   * A human readable comment describing the table.
+   * @type {String}
+   */
+  this.comment = null;
+  /**
+   * Specifies the time to wait before garbage collecting tombstones (deletion markers)
+   * @type {number}
+   */
+  this.gcGraceSeconds = 0;
+  /**
+   * Compaction strategy class used for the table.
+   * @type {String}
+   */
+  this.compactionClass = null;
+  /**
+   * Associative-array containing the compaction options keys and values.
+   * @type {Object}
+   */
+  this.compactionOptions = null;
+  /**
+   * Associative-array containing the compression options.
+   * @type {Object}
+   */
+  this.compression = null;
+  /**
+   * Specifies the probability of read repairs being invoked over all replicas in the current data center.
+   * @type {number}
+   */
+  this.localReadRepairChance = 0;
+  /**
+   * Specifies the probability with which read repairs should be invoked on non-quorum reads. The value must be between 0 and 1.
+   * @type {number}
+   */
+  this.readRepairChance = 0;
+  /**
+   * Applies only to counter tables.
+   * When set to true, replicates writes to all affected replicas regardless of the consistency level specified by
+   * the client for a write request. For counter tables, this should always be set to true.
+   * @type {boolean}
+   */
+  this.replicateOnWrite = true;
+  /**
+   * Determines  whether the table uses the COMPACT STORAGE option.
+   * @type {boolean}
+   */
+  this.isCompact = false;
+  /**
+   * Array describing the table columns.
+   * @type {Array}
+   */
+  this.columns = [];
+  /**
+   * An associative Array of columns by name.
+   * @type {Object}
+   */
+  this.columnsByName = null;
+  /**
+   * Array describing the columns that are part of the partition key.
+   * @type {Array}
+   */
+  this.partitionKeys = [];
+  /**
+   * Array describing the columns that form the clustering key.
+   * @type {Array}
+   */
+  this.clusteringKeys = [];
+  /**
+   * Array describing the clustering order of the columns in the same order as the clusteringKeys.
+   * @type {Array}
+   */
+  this.clusteringOrder = [];
+}
+
+util.inherits(TableMetadata, events.EventEmitter);
+
+module.exports = TableMetadata;
